requirements.txt
Django
djangorestframework
django-froala-editor